Preliminay maps support.
git-svn-id: https://klub.com.pl/var/svn/omnet@10 056025b2-ba0e-0410-831a-f040b6dfa49e

+++ wimax/wmaxmsg.h
@@ -5,8 +5,8 @@
 
 /* Information Element (simplified version) */
 struct WMaxMapIE {
-    int cid;
-    int dataLen;
+    uint16_t cid;
+    uint32_t dataLen;
 };
